DatabaseDriverFactory for the common package and it's implementation for both Android and iOS
DatabaseDriverFactory for the common package and its implementation for both Android and iOS.

An error starts showing at this level when building the app targeting iOS:
"ld: symbol(s) not found for architecture x86_64"
It seams there are limitations to which target architectures KMM can build!
This issue has taken so much time, on multiple sample projects I've worked on and I couldn't find a solution so far!